refactor: rename files to denote the new client naming scheme
