+++ resources/ansible/roles/client-metrics/tasks/main.yml
