+++ resources/terraform/client/digital-ocean/main.tf
